+++ client/src/components/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 // import { hot } from 'react-hot-loader/root';
 import StarRating from "./Overview/ProductInformation/StarRating.jsx"
 import ReviewWidget from './R&R/ReviewWidget.jsx';
-import Card from './components/Card.jsx';
+import Carousel from './Related/Carousel.jsx';
 
 
 
@@ -11,8 +11,8 @@ class App extends React.Component {
    // const { name } = 'Robert';
     return (
       <>
-        < StarRating />
-        <Card />
+        <StarRating />
+        <Carousel />
         {/* <ReviewWidget /> */}
 
      </>
