+++ client/src/components/Related/Carousel.jsx
@@ -1,0 +1,15 @@
+import React, { useState, useEffect } from 'react';
+import Card from './Card.jsx';
+
+var Carousel = () => {
+
+
+
+  return (
+    <div className="carousel">
+      <Card />
+    </div>
+  )
+}
+
+export default Carousel = Carousel;
