--- client/src/components/Related/config.js
@@ -1,6 +1,0 @@
-let config = {
-  TOKEN: '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',
-  API_URL: 'https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfp'
-}
-
-export default config = config;
